requirements.txt
bkcharts==0.2
bleach==2.0.0
bokeh==0.12.6
certifi==2017.4.17
chardet==3.0.4
cycler==0.10.0
Cython==0.25.2
dateutils==0.6.6
decorator==4.0.11
emoji==0.4.5
entrypoints==0.2.3
html5lib==0.999999999
idna==2.5
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
llvmlite==0.18.0
MarkupSafe==1.0
Mastodon.py==1.0.8
matplotlib==2.0.2
mistune==0.7.4
mpmath==0.19
natto-py==0.8.0
nbconvert==5.2.1
nbformat==4.3.0
networkx==1.11
notebook==5.0.0
numba==0.33.0
numpy==1.13.0
olefile==0.44
pandas==0.20.2
pandocfilters==1.4.1
patsy==0.4.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
pivottablejs==0.7.0
prompt-toolkit==1.0.14
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.18.1
scikit-image==0.13.0
scikit-learn==0.18.2
scipy==0.19.1
seaborn==0.7.1
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.11
sympy==1.0
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
wordcloud==1.3.1
